test(cli): move tests into __tests__ folder per repo convention
